pyproject.toml
[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[project]
name = "comfy_openai_image_api_azure"
version = "0.0.4"
description = "A ComfyUI node that provides access to OpenAI's image generation and editing capabilities, including support for gpt-image-1 model with both OpenAI and Azure OpenAI providers."
authors = [
  {name = "Jiajun Chen", email = "cjj198909@github.com"},
  {name = "Xin (Original Author)", email = "unicough.github@gmail.com"}
]
readme = "README.md"
license = {text = "MIT license"}
classifiers = []
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "numpy>=1.21.0",
    "requests>=2.25.0",
]

[project.optional-dependencies]
dev = [
    "bump-my-version",
    "coverage",  # testing
    "mypy",  # linting
    "pre-commit", # runs linting on commit
    "pytest",  # testing
    "ruff",  # linting
]

[project.urls]
bugs = "https://github.com/cjj198909/comfy_openai_image_api_azure/issues"
homepage = "https://github.com/cjj198909/comfy_openai_image_api_azure"
Repository = "https://github.com/cjj198909/comfy_openai_image_api_azure.git"
Original = "https://github.com/unicough/comfy_openai_image_api"


[tool.comfy]
PublisherId = "unicough"
DisplayName = "OpenAI/Azure OpenAI Image API"
Icon = ""

[tool.setuptools.package-data]
"*" = ["*.*"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = [
    "tests",
]

[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
allow_untyped_defs = true
disable_error_code = "attr-defined"

[tool.ruff]
# extend-exclude = ["static", "ci/templates"]
line-length = 140
src = ["src", "tests"]
target-version = "py39"

# Add rules to ban exec/eval
[tool.ruff.lint]
select = [
    "S102",  # exec-builtin
    "S307",  # eval-used
    "W293",
    "F",  # The "F" series in Ruff stands for "Pyflakes" rules, which catch various Python syntax errors and undefined names.
    # See all rules here: https://docs.astral.sh/ruff/rules/#pyflakes-f
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
